Create Siria
Accepts a file with a list of countries and their population, then takes another file from the console that also has a list of countries and the number of fighters sent to Syria from that country. The program then outputs a list of countries with the quotient of fighters per country sent to Syria (fighters/total_population).