Create BES
This program takes in a file with a list of companies and the investment value made by the bank BES on each company. Then takes another file from the console that has a list of companies and another set of values of investment made by BES. The final output gives us the differences found in both files when the same company comes up in both. The goal is to tells by how much did BES lie in the investments (one file is supposed to be the public list the bank published and the other the internal true list of investments).